Fix bad vegs bug

+++ Assets/_Project/Scripts/Configs/VegetableConfig.cs
@@ -5,7 +5,7 @@ namespace _Project.Scripts.Configs
     public abstract class VegetableConfig : ScriptableObject
     {
         [field: SerializeField] public string Name { get; private set; }
-        [field: SerializeField] public int Score { get; private set; }
+        [field: SerializeField] public float Score { get; private set; }
         [field: SerializeField] public float Speed { get; private set; }
         [field: SerializeField] public bool IsGood { get; private set; }
     }
